api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
PyMuPDF==1.24.5
litellm==1.40.14
python-dotenv==1.0.1
